bike_analysis/source/cloud_functions/station_inventory/requirements.txt
db-dtypes==1.2.0
google-cloud-bigquery==3.17.1
pymsteams==0.2.2